#2860: Optimized OrderByDropdown component
