+++ roles/lib/files/FWO.Basics/StringExtensionsIp.cs
@@ -196,6 +196,18 @@ namespace FWO.Basics
             return bytes.Take(targetLength).ToArray();  // Ensure it's exactly targetLength bytes
         }
 
+        /// <summary>
+        /// Parses a string to an IPAdress object and a string that represents the subnet mask (empty if there is no subnet mask).
+        /// </summary>
+        public static (IPAddress, string) ToIPAdressAndSubnetMask(this string str)
+        {
+            IPAddress ipAdress = IPAddress.Parse(str.StripOffNetmask());
+            string subnetMask = "";
+            var hasSubnetMask = str.TryGetNetmask(out subnetMask);
+
+            return (ipAdress, hasSubnetMask ? subnetMask.Substring(1) : "");
+        }
+
     }
 }
