+++ roles/ui/files/FWO.UI/Data/OrderMode.cs
@@ -1,0 +1,9 @@
+namespace FWO.Ui.Data
+{
+    public enum OrderMode
+    {
+        Asc,
+        Desc        
+    }
+
+}
